// frontend/src/types/api.generated.ts
export interface paths {
    "/applications": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: {
            parameters: {
                query?: {
                    "page-number"?: number;
                    "page-size"?: number;
                    "search-term"?: string;
                    "sort-by"?: string;
                    "sort-direction"?: string;
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Success */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        put?: never;
        post: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: {
                content: {
                    "application/json": components["schemas"]["ApplicationRequest"];
                    "text/json": components["schemas"]["ApplicationRequest"];
                    "application/*+json": components["schemas"]["ApplicationRequest"];
                };
            };
            responses: {
                /** @description Success */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
// ...
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}